renaming supabase_session
