+++ src/leapfrogai_api/routers/openai/chat.py
@@ -13,7 +13,7 @@ from leapfrogai_api.backend.grpc_client import (
 )
 from leapfrogai_api.backend.helpers import grpc_chat_role
 from leapfrogai_api.backend.types import ChatCompletionRequest
-from leapfrogai_api.routers.supabase_session_api_key import Session
+from leapfrogai_api.routers.supabase_session import Session
 from leapfrogai_api.utils import get_model_config
 from leapfrogai_api.utils.config import Config
 from leapfrogai_sdk.chat.chat_pb2 import (
