+++ src/leapfrogai_api/routers/openai/requests/create_message_request.py
@@ -8,7 +8,7 @@ from openai.types.beta.threads import MessageContent, TextContentBlock, Text, Me
 from openai.types.beta.threads.message import Attachment
 from pydantic import BaseModel, Field
 from leapfrogai_api.data.crud_message import CRUDMessage
-from leapfrogai_api.routers.supabase_session_api_key import Session
+from leapfrogai_api.routers.supabase_session import Session
 
 
 class CreateMessageRequest(BaseModel):
