+++ src/leapfrogai_api/routers/openai/threads.py
@@ -8,7 +8,7 @@ from leapfrogai_api.routers.openai.requests.create_thread_request import (
     CreateThreadRequest,
 )
 from leapfrogai_api.data.crud_thread import CRUDThread
-from leapfrogai_api.routers.supabase_session_api_key import Session
+from leapfrogai_api.routers.supabase_session import Session
 from leapfrogai_api.utils.validate_tools import (
     validate_tool_resources,
 )
