+++ src/leapfrogai_api/routers/openai/vector_stores.py
@@ -23,7 +23,7 @@ from leapfrogai_api.data.crud_vector_store_file import (
     CRUDVectorStoreFile,
     FilterVectorStoreFile,
 )
-from leapfrogai_api.routers.supabase_session_api_key import Session
+from leapfrogai_api.routers.supabase_session import Session
 
 router = APIRouter(prefix="/openai/v1/vector_stores", tags=["openai/vector_stores"])
 security = HTTPBearer()
